如果项目根目录有composer.phar，则运行此文件，否则运行全局Composer命令

+++ composer_install.sh
@@ -5,4 +5,11 @@ all_version=($(ls ./deploy/$1 -t))
 last_version=${all_version[0]}
 
 cd ./deploy/$1/${last_version}
-composer install -vvv 2>&1
+
+if [ ! -f "./composer.phar" ]; then
+  echo "Running Global Composer"
+  composer install -vvv 2>&1
+else
+  echo "Running Project Local Composer.phar"
+  php composer.phar install -vvv 2>&1
+fi
